+++ internal/storage/storage.go
@@ -3,6 +3,7 @@ package storage
 import (
 	"errors"
 	"fmt"
+	"io"
 	"time"
 )
 
@@ -45,4 +46,19 @@ type ServerError struct {
 
 func (s *ServerError) Error() string {
 	return fmt.Sprintf("server error with status '%d'; title '%s'; msg '%s'", s.Code, s.Title, s.Msg)
+}
+
+// AppService is the interface for interacting with the Sauce application storage.
+type AppService interface {
+	Upload(name string) (ArtifactMeta, error)
+	// UploadStream uploads the contents of reader and stores them under the given filename.
+	UploadStream(filename string, reader io.Reader) (ArtifactMeta, error)
+	Download(id string) (io.ReadCloser, int64, error)
+	Find(name string) (ArtifactMeta, error)
+	List(opts ListOptions) (List, error)
+}
+
+// ArtifactMeta represents metadata of the uploaded file.
+type ArtifactMeta struct {
+	ID string
 }
